+++ src/app/Http/Controllers/CrudController.php
@@ -15,11 +15,12 @@ use Backpack\CRUD\app\Http\Controllers\CrudFeatures\AjaxTable;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Reorder;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Revisions;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\ShowDetailsRow;
+use Backpack\CRUD\app\Http\Controllers\CrudFeatures\SaveActions;
 
 class CrudController extends BaseController
 {
     use DispatchesJobs, ValidatesRequests;
-    use AjaxTable, Reorder, Revisions, ShowDetailsRow;
+    use AjaxTable, Reorder, Revisions, ShowDetailsRow, SaveActions;
 
     public $data = [];
     public $crud;
@@ -82,6 +83,7 @@ class CrudController extends BaseController
 
         // prepare the fields you need to show
         $this->data['crud'] = $this->crud;
+        $this->data['saveAction'] = $this->getSaveAction();
         $this->data['fields'] = $this->crud->getCreateFields();
         $this->data['title'] = trans('backpack::crud.add').' '.$this->crud->entity_name;
 
@@ -118,14 +120,10 @@ class CrudController extends BaseController
         // show a success message
         \Alert::success(trans('backpack::crud.insert_success'))->flash();
 
-        // redirect the user where he chose to be redirected
-        switch ($request->input('redirect_after_save')) {
-            case 'current_item_edit':
-                return \Redirect::to($this->crud->route.'/'.$item->getKey().'/edit');
+        // save the redirect choice for next time
+        $this->setSaveAction();
 
-            default:
-                return \Redirect::to($request->input('redirect_after_save'));
-        }
+        return $this->performSaveAction($item->getKey());
     }
 
     /**
@@ -142,6 +140,7 @@ class CrudController extends BaseController
         // get the info for that entry
         $this->data['entry'] = $this->crud->getEntry($id);
         $this->data['crud'] = $this->crud;
+        $this->data['saveAction'] = $this->getSaveAction();
         $this->data['fields'] = $this->crud->getUpdateFields($id);
         $this->data['title'] = trans('backpack::crud.edit').' '.$this->crud->entity_name;
 
@@ -181,7 +180,10 @@ class CrudController extends BaseController
         // show a success message
         \Alert::success(trans('backpack::crud.update_success'))->flash();
 
-        return \Redirect::to($this->crud->route);
+        // save the redirect choice for next time
+        $this->setSaveAction();
+
+        return $this->performSaveAction();
     }
 
     /**