+++ src/resources/lang/it/crud.php
@@ -13,17 +13,17 @@ return [
     |
     */
 
+    // Forms
+    'save_action_save_and_new' => 'Save and new item', // TODO: translate
+    'save_action_save_and_edit' => 'Save and edit this item', // TODO: translate
+    'save_action_save_and_back' => 'Save and back', // TODO: translate
+    'save_action_changed_notification' => 'Default save action changed', // TODO: translate
+
     // Create form
     'add'                 => 'Aggiungi',
     'back_to_all'         => 'Torna a tutti i ',
     'cancel'              => 'Annulla',
     'add_a_new'           => 'Aggiungi un nuovo ',
-
-        // Create form - advanced options
-        'after_saving'            => 'Dopo il salvataggio',
-        'go_to_the_table_view'    => 'vai alla lista di tutti gli elementi',
-        'let_me_add_another_item' => 'aggiungi un altro elemento',
-        'edit_the_new_item'       => 'modifica il nuovo elemento',
 
     // Edit form
     'edit'                 => 'Modifica',