+++ src/resources/views/edit.blade.php
@@ -34,10 +34,11 @@
 		      	@include('crud::form_content', ['fields' => $crud->getFields('update', $entry->getKey())])
 		      @endif
 		    </div><!-- /.box-body -->
-		    <div class="box-footer">
 
-			  <button type="submit" class="btn btn-success ladda-button" data-style="zoom-in"><span class="ladda-label"><i class="fa fa-save"></i> {{ trans('backpack::crud.save') }}</span></button>
-		      <a href="{{ url($crud->route) }}" class="btn btn-default ladda-button" data-style="zoom-in"><span class="ladda-label">{{ trans('backpack::crud.cancel') }}</span></a>
+            <div class="box-footer">
+
+                @include('crud::inc.form_save_buttons')
+
 		    </div><!-- /.box-footer-->
 		  </div><!-- /.box -->
 		  {!! Form::close() !!}