+++ src/resources/views/form_content.blade.php
@@ -38,6 +38,17 @@
 	<script>
         jQuery('document').ready(function($){
 
+          //Save button dropdown toggles
+          var saveActions = $('#saveActions'),
+          crudForm        = saveActions.parents('form'),
+          saveActionField = $('[name="save_action"]');
+
+          saveActions.on('click', '.dropdown-menu a', function(){
+              var saveAction = $(this).data('value');
+              saveActionField.val( saveAction );
+              crudForm.submit();
+          });
+
       		// Ctrl+S and Cmd+S trigger Save button click
       		$(document).keydown(function(e) {
       		    if ((e.which == '115' || e.which == '83' ) && (e.ctrlKey || e.metaKey))