+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All Notable changes to `Backpack CRUD` will be documented in this file
 
+-----------
+
+## 3.2.0 - YYYY-MM-DD
+
+### Added
+- form save button better UI&UX: they have the options in a dropdown instead of radio buttons and the default behaviour is stored in the session upon change - thanks to [Owen Melbourne](https://github.com/OwenMelbz);
+
+-----------
+
 ## NEXT - YYYY-MM-DD
 
 ### Added
@@ -18,6 +27,19 @@ All Notable changes to `Backpack CRUD` will be documented in this file
 
 ### Security
 - Nothing
+
+
+## [3.2.0] - 2016-11-13
+
+### Added
+- redirect_after_save button actions;
+- filters on list views (deleted the 3.1.41 and 4.1.42 tags because they were breaking changes);
+
+
+## [3.1.42] - 2016-11-13
+
+### Fixed
+- n-n filters prevented CRUD items from being added;
 
 
 ## [3.1.41] - 2016-11-11